% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCanStatF}
\alias{getCanStatF}
\title{Function to calculate the canonical statistic.}
\usage{
getCanStatF(Bds, Cols, ncolors)
}
\arguments{
\item{Bds}{matrix of bonds}

\item{Cols}{Vector of colors.}

\item{ncolor}{number of spins.}
}
\description{
Function to calculate the canonical statistic.
}
\examples{
Obs.loc = as.matrix(expand.grid(1:4,1:4))
Bds = getBonds(Obs.loc, NN = 4)
Bds = Bds[which(Bds[,1]>0),]
Betas = c(0.8, 0.8, 0.2)
Cols = sample(1:3, 16, replace = TRUE)
par(mfrow = c(1,1))
col = grey.colors(3)
image(matrix(Cols, ncol = 4), col = col)
CS = getCanStatF(Bds, Cols, 3)
}
