% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loopSW}
\alias{loopSW}
\title{Function performing the loop in the SW algorithm.}
\usage{
loopSW(Bds, Cols, ncolors, Nrun, Betas)
}
\arguments{
\item{Bds}{matrix of bonds}

\item{Cols}{Vector of colors.}

\item{Nrun}{number of runs.}

\item{Betas}{vector of parameters.}

\item{ncolor}{number of spins.}
}
\description{
Function performing the loop in the SW algorithm.
}
\examples{
Obs.loc = as.matrix(expand.grid(1:10,1:10))
Bds = getBonds(Obs.loc, NN = 4)
Betas = c(0.8, 0.8, 0.2)
Cols = sample(1:3, 100, replace = TRUE)
par(mfrow = c(1,1))
col = grey.colors(3)
image(matrix(Cols, ncol = 10), col = col)
CS = loopSW(Bds, Cols, 3, 1000, Betas)
image(matrix(CS, ncol = 10), col = col)
}
