% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{row_erase}
\alias{row_erase}
\title{Function to suppress rows from a matrix}
\usage{
row_erase(x, rowID)
}
\arguments{
\item{x}{matrix which rows need to be erased}

\item{rowID}{row index}
}
\description{
Function to suppress rows from a matrix
}
\examples{
x<- matrix(1:9, ncol = 3)
row_erase(x,1:2)
}
