% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{selectBonds}
\alias{selectBonds}
\title{Function to select bonds based on colors and parameters}
\usage{
selectBonds(Bds, Cols, Betas)
}
\arguments{
\item{Bds}{matrix of neighbouring bonds.}

\item{Cols}{vector of colors per vertex.}

\item{Betas}{vector of parameters for the Potts model.}
}
\description{
Function to select bonds based on colors and parameters
}
\examples{
Obs.loc = expand.grid(1:10,1:10)
Bds = getBonds(Obs.loc)
Bds = Bds[which(Bds[,1]>0),]
Betas = c(0.8, 0.8, 0.2)
Cols = sample(1:3, 100, replace = TRUE)
Bd = selectBonds(Bds, Cols, Betas)
par(mfrow = c(1,1))
col = grey.colors(3)
image(matrix(Cols, ncol = 10), col = col)
segments((Obs.loc[Bd[,1],1]-1)/9, (Obs.loc[Bd[,1],2]-1)/9, (Obs.loc[Bd[,2],1]-1)/9, (Obs.loc[Bd[,2],2]-1)/9)
segments((Obs.loc[,1]-1)/9, (Obs.loc[,2]-1)/9, (Obs.loc[,1]-1)/9, (Obs.loc[,2]-1)/9, lwd = 4)
}
